requirements.txt
quart==0.19.6
werkzeug==3.0.4
nltk
scikit-learn
openai==0.28.0
aiohttp==3.10.5
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
aiofiles==24.1.0
aiosqlite==0.20.0
rapidfuzz==3.10.1